+++ Pages/RendererV2.cshtml.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace CardConjurer.Pages;
+
+public class RendererV2Model : PageModel
+{
+    public void OnGet()
+    {
+    }
+}
+